requirements.txt
blockfrost-python==0.5.3
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
numpy==1.25.0
pandas==1.3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
urllib3==2.0.4
